requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pdf2image
Pillow
litellm
boto3
weaviate-client
aiofiles
